# docs/superpowers/specs/monexus-api-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MoNexus API",
    "version": "1.9.0",
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发，所有业务路径带 /api 前缀"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/payment/webhooks/stripe": {
      "post": {
        "tags": [
          "Payment Webhooks"
        ],
        "summary": "stripe 异步通知",
        "security": [],
        "description": "Raw body 验签。成功后先持久化 observation 再异步 apply。不使用 userId/orderId 作为指标标签。",
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "渠道要求的成功 ACK"
          },
          "400": {
            "description": "验签失败"
          },
          "503": {
            "description": "adapter 未注册"
          }
        }
      }
    },
    "/api/payment/webhooks/paypal": {
      "post": {
        "tags": [
          "Payment Webhooks"
        ],
        "summary": "paypal 异步通知",
        "security": [],
        "description": "Raw body 验签。成功后先持久化 observation 再异步 apply。不使用 userId/orderId 作为指标标签。",
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "渠道要求的成功 ACK"
          },
          "400": {
            "description": "验签失败"
          },
          "503": {
            "description": "adapter 未注册"
          }
        }
      }
    },
    "/api/payment/webhooks/wechat-pay": {
      "post": {
        "tags": [
          "Payment Webhooks"
        ],
        "summary": "wechat_pay 异步通知",
        "security": [],
        "description": "Raw body 验签。成功后先持久化 observation 再异步 apply。不使用 userId/orderId 作为指标标签。",
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "渠道要求的成功 ACK"
          },
          "400": {
            "description": "验签失败"
          },
          "503": {
            "description": "adapter 未注册"
          }
        }
      }
    },
    "/api/payment/webhooks/alipay": {
      "post": {
        "tags": [
          "Payment Webhooks"
        ],
        "summary": "alipay 异步通知",
        "security": [],
        "description": "Raw body 验签。成功后先持久化 observation 再异步 apply。不使用 userId/orderId 作为指标标签。",
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "渠道要求的成功 ACK"
          },
          "400": {
            "description": "验签失败"
          },
          "503": {
            "description": "adapter 未注册"
          }
        }
      }
    },
    "/api/payment/webhooks/simulator": {
      "post": {
        "tags": [
          "Payment Webhooks"
        ],
        "summary": "simulator 异步通知",
        "security": [],
        "description": "Raw body 验签。成功后先持久化 observation 再异步 apply。不使用 userId/orderId 作为指标标签。",
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "渠道要求的成功 ACK"
          },
          "400": {
            "description": "验签失败"
          },
          "503": {
            "description": "adapter 未注册"
          }
        }
      }
    },
...
  }
}